Rename prompts.go -> prompt_types.go, prompts_generic.go -> prompts.go
The previous naming was backwards: cmd/prompts.go was actually the
shared factory used by routines, policies, glossaries, variables, and
macros, while cmd/prompts_generic.go was the specific iai prompts
command for plain prompts. Swap them so the filename reflects the
contents.
